+++ src/irmin-mirage/git/irmin_mirage_git_intf.ml
@@ -20,7 +20,7 @@ module type S = sig
       with type Private.Remote.endpoint = Mimic.ctx * Smart_git.Endpoint.t
 
   val remote :
-    ?ctx:Mimic.ctx -> ?headers:Cohttp.Header.t -> string -> Irmin.remote
+    ?ctx:Mimic.ctx -> ?headers:(string * string) list -> string -> Irmin.remote
 end
 
 module type Maker = sig
@@ -67,7 +67,7 @@ module type KV_RO = sig
     ?branch:string ->
     ?root:key ->
     ?ctx:Mimic.ctx ->
-    ?headers:Cohttp.Header.t ->
+    ?headers:(string * string) list ->
     git ->
     string ->
     t Lwt.t
@@ -87,7 +87,7 @@ module type KV_RW = sig
     ?branch:string ->
     ?root:key ->
     ?ctx:Mimic.ctx ->
-    ?headers:Cohttp.Header.t ->
+    ?headers:(string * string) list ->
     ?author:(unit -> string) ->
     ?msg:([ `Set of key | `Remove of key | `Batch ] -> string) ->
     git ->